[Update financial profile] Update schema
